seperate parser from code, verify this is working -- then to add ipstack
