+++ go.mod
@@ -4,5 +4,6 @@ go 1.13
 
 require (
 	github.com/oschwald/geoip2-golang v1.5.0
+	github.com/qioalice/ipstack v1.0.1 // indirect
 	golang.org/x/sys v0.0.0-20210223212115-eede4237b368 // indirect
 )
